# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Digital Portfolio API
  description: API for contact and AI assistant endpoints
  version: 1.0.0
servers:
  - url: https://your-deployment.example.com
paths:
  /api/contact:
    post:
      summary: Submit contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                message:
                  type: string
                company:
                  type: string
              required:
                - name
                - email
                - message
      responses:
        '200':
          description: Message delivered
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          description: Validation error
  /api/ai-chat:
    post:
      summary: AI Chat endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                conversationId:
                  type: string
              required:
                - message
      responses:
        '200':
          description: AI response
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                  conversationId:
                    type: string
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
components:
  securitySchemes: {}